// src/shared/types/generated/application/SourceCatalogEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SourceCatalogEntry = { id: string, name: string, domain: string, credential_owner_site_id: string, example_query: string, supports_query: boolean, supports_account: boolean, auth_required_for_full_access: boolean, auth_strictly_required: boolean, credential_types: Array<string>, oauth_provider: string | null, };