// src/shared/types/generated/application/SubscriptionRunActivity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SubscriptionQueryActivity } from "./SubscriptionQueryActivity";
import type { SubscriptionRunSummary } from "./SubscriptionRunSummary";

export type SubscriptionRunActivity = { summary: SubscriptionRunSummary, queries: Array<SubscriptionQueryActivity>, };