// src/shared/types/generated/application/CurrentSubscriptionProgress.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActivityCounts } from "./ActivityCounts";

export type CurrentSubscriptionProgress = { subscription_id: number, run_id: number, status: string, requested_by: string, counts: ActivityCounts, gallery_total_items: number | null, };