// src/shared/types/generated/application/NewSubscription.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NewSubscriptionQuery } from "./NewSubscriptionQuery";

export type NewSubscription = { name: string, schedule: string, initial_post_limit: number | null, periodic_post_limit: number | null, queries: Array<NewSubscriptionQuery>, };