// src/shared/types/generated/application/SubscriptionCoverCandidatePage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SubscriptionCoverCandidate } from "./SubscriptionCoverCandidate";
import type { SubscriptionCoverCandidateCursor } from "./SubscriptionCoverCandidateCursor";

export type SubscriptionCoverCandidatePage = { candidates: Array<SubscriptionCoverCandidate>, next_cursor: SubscriptionCoverCandidateCursor | null, };